move stories to component folders
