+++ .storybook/config.js
@@ -1,15 +1,13 @@
 import { configure } from '@storybook/react'
 
-const req = require.context('../stories', true, /\.stories\.js$/);
+import '../src/styles/core.scss'
+import '../src/styles/icons.scss'
+import '../src/styles/storybook.scss'
+
+const req = require.context('../src', true, /\.stories\.js$/);
 
 function loadStories() {
-  require('../stories/index.js')
   req.keys().forEach(filename => req(filename));
 }
-
-// function loadStories() {
-//   require('../stories/index.js')
-//   // You can require as many stories as you need.
-// }
 
 configure(loadStories, module)
