+++ src/components/AssetItem/AssetItem.stories.js
