+++ src/components/base/JIcon/JIcon.stories.js
@@ -4,16 +4,16 @@ import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { keyBy } from 'lodash-es'
 
-import JIcon from '../../../src/components/base/JIcon'
+import JIcon from './JIcon'
 
 const filesSmall = require.context(
-  '../../../src/public/assets/icons/sprite-pack/small', true, /.*\.svg$/)
+  '../../../public/assets/icons/sprite-pack/small', true, /.*\.svg$/)
 const filesMedium = require.context(
-  '../../../src/public/assets/icons/sprite-pack/medium', true, /.*\.svg$/)
+  '../../../public/assets/icons/sprite-pack/medium', true, /.*\.svg$/)
 const filesLarge = require.context(
-  '../../../src/public/assets/icons/sprite-pack/large', true, /.*\.svg$/)
+  '../../../public/assets/icons/sprite-pack/large', true, /.*\.svg$/)
 const filesXLarge = require.context(
-  '../../../src/public/assets/icons/sprite-pack/xlarge', true, /.*\.svg$/)
+  '../../../public/assets/icons/sprite-pack/xlarge', true, /.*\.svg$/)
 
 const iconsSmall = keyBy(
   filesSmall.keys().map(x => filesSmall(x).default),
