+++ src/components/base/JLoader/JLoader.stories.js
@@ -3,7 +3,7 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import JLoader from '../../../src/components/base/JLoader'
+import JLoader from './JLoader'
 
 storiesOf('JLoader')
   .add('Different colors', () => (
@@ -17,7 +17,11 @@ storiesOf('JLoader')
         <JLoader color='gray' />
       </div>
       <h2>white</h2>
-      <div style={{ padding: '20px', backgroundColor: '#666' }}>
+      <div style={{
+        padding: '20px',
+        backgroundColor: '#666',
+      }}
+      >
         <JLoader color='white' />
       </div>
     </div>
