--- stories/index.js
@@ -1,8 +1,0 @@
-// @flow
-
-import '../src/styles/core.scss'
-import '../src/styles/icons.scss'
-import './styles.scss'
-
-// import './components/base'
-// import './components'
